Update client.rb
Line 126 references the instance variable self.smb.read_timeout. SMB is an instance of simpleclient (Lines 118 || 120) and does not have an instance variable by that name. Client, which is an instance variable of a simpleclient object, does. I updated this line to access the client instance variable. Testing this appears to work as intended.